Offer to give a talk
I have a talk prepared that I'll be giving in late May at DevOpsDays Toronto on the way we're using chat automation. I could present this talk on May 10 at DevOpsDC as a practice run, or at a later meeting once I figure out which jokes land and which should be cut.